// src/schema/enums/AUTH_ERROR_ENUM.ts
export enum AUTH_ERROR_ENUM {
    TOKEN_FORGED = "TOKEN_FORGED",
    TOKEN_EXPIRED = "TOKEN_EXPIRED",
    TOKEN_DESTROYED = "TOKEN_DESTROYED",
    TOKEN_INVALID = "TOKEN_INVALID",
    SECRET_INVALID = "SECRET_INVALID",
    TOKEN_NOT_AUTHORIZED = "TOKEN_NOT_AUTHORIZED",
    IP_NOT_CORRESPONDING = "IP_NOT_CORRESPONDING",
    UA_NOT_CORRESPONDING = "UA_NOT_CORRESPONDING",
}